Logica TP6 teoricamente finalizada. - La Ec Diferencial de la duración de bloqueo de llegadas no funciona, da siempre 0. Lo demás en teoria funciona bien.

+++ SimulacionesTp6/src/main/java/p/grupo/k1/simulacionestp6/modelo/rungeKutta/ResultadoRungeKutta.java
@@ -11,14 +11,14 @@ import lombok.Data;
 @Data
 public class ResultadoRungeKutta {
 
-    private float xm;
-    private float ym;
-    private float xmp1;
-    private float k1;
-    private float k2;
-    private float k3;
-    private float k4;
-    private float ymp1;
+    private double xm;
+    private double ym;
+    private double xmp1;
+    private double k1;
+    private double k2;
+    private double k3;
+    private double k4;
+    private double ymp1;
 
 
 }
